[REFACTOR] Update react structure about post
