+++ client/src/components/PostView/Post/PostLike/PostLike.style.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import styled from 'styled-components';
+import heartFill from '../../../../asset/icon/icon-heart-fill.svg';
+import heartLine from '../../../../asset/icon/icon-heart-line.svg';
+
+export default function PostLikePresenter({ handleLikeClick, isLikeChecked, likeValue }) {
+  return (
+    <PostLikeSpan onClick={handleLikeClick} likechecked={isLikeChecked}>
+      {likeValue}
+    </PostLikeSpan>
+  );
+}
+
+const PostLikeSpan = styled.span`
+  display: flex;
+  margin-right: 12px;
+  cursor: pointer;
+  &::before {
+    content: '';
+    width: 20px;
+    height: 20px;
+    display: block;
+    background: ${(props) => (props.likechecked ? `url(${heartFill})` : `url(${heartLine})`)};
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: contain;
+    margin-right: 7px;
+  }
+`;
